release: v0.16.0 -- 7-phase refactor complete + post-release followups (#48)
* feat(release): v0.16.0 -- full 7-phase refactor with alembic round-trip fix

Phases 1-7 deliver a complete structural overhaul of the backend
and a frontend polish pass:

Phase 1 (DB + Security)
  - composite indexes on fight_agents(account_name, fight_id),
    webhook_deliveries(subscription_id, next_attempt_at),
    guild_members(account_name)
  - 30/minute rate limiting on players, webhooks, guilds, skills
  - FK webhook_dlq.subscription_id -> webhook_subscriptions.id
  - /healthz S3 + DB connectivity checks
  - webhook HTTP timeout configurable via WEBHOOK_DISPATCH_TIMEOUT_S
  - MIME validation on upload
  - KEK loaded-from-.env warning at Settings load
  - optional require_auth decorator + API_KEY config field

Phase 2 (Repository + Service Layer)
  - NEW repositories/ package: FightRepository, UploadRepository,
    WebhookRepository, PlayerRepository, GuildRepository
  - 6 services refactored to delegate queries to repos
  - routes/players.py slimmed 882 -> 231 lines (-74%)
  - 5 ADR documents in docs/adr/ explaining the pattern

Phase 3 (Schema Normalization)
  - NEW OrmFightPlayerBoon model with PK (fight_id, account_name,
    boon_name) dual-written from player_summaries.py
  - INTEGER -> BIGINT for total_damage/healing/buff_removal,
    power_damage, condi_damage, uploads.size_bytes
  - TimestampMixin adds created_at+updated_at to fights, fight_agents,
    fight_skills
  - 9 SCAFFOLD / transformation-output columns dropped explicitly:
    fight_player_summaries.{deaths, stun_breaks, interrupts, downs,
    dodges, damage_taken, blocked} (Phase 6 v2 SCAFFOLD-zero
    contract retired at v0.12.3), fights.context,
    fight_agents.position_samples. All 9 are restored on
    downgrade -1; see CHANGELOG Phase 3 destructive drops.

Phase 4 (Performance)
  - streaming upload for files >10 MB to tempfile with incremental SHA-256
  - iterator-based parsing (removed list() on _parser.parse())
  - streaming JSONL+gzip via BytesIO + GzipFile.writelines
  - keyset pagination on /api/v1/players via PlayerRepository
    (X-Next-Cursor header)
  - NEW CacheService (services/cache_service.py) with Redis backend
  - batch UPSERT via postgres_insert.on_conflict_do_update

Phase 5 (Quality)
  - _persist_player_summaries split into _process_events_to_buckets
    + _write_summary_and_boon_rows
  - dead code removal (sync_guilds stub, port=1 guard, redundant
    crypto/schema_guard docstrings)
  - strict types: combine_day_midnight signature tightened
  - NEW .github/workflows/security.yml (Trivy fs scan on every PR)
  - NEW detect-secrets pre-commit hook + .secrets.baseline allowlist

Phase 6 (Monitoring)
  - NEW RequestIDMiddleware injecting X-Request-Id into every
    response and request.state.request_id for log correlation

Phase 7 (Documentation)
  - CONTRIBUTING.md layered architecture section (Repository ->
    Service -> Route dependency inversion + per-layer conventions)
  - 5 ADRs in docs/adr/: 001-repository-pattern, 002-service-layer,
    003-boon-normalization, 004-keyset-pagination,
    005-streaming-performance

ALEAMBIC CHAIN SAFETY FIX
  - 0018_player_boon_uptimes.downgrade() now uses
    op.drop_column(if_exists=True) for both _BOON_UPTIME_COLUMNS
    and _OUTGOING_COLUMNS loops. Without IF EXISTS, the chain
    reversal past 0018 raised UndefinedColumn on outgoing_stealth
    (and 13 sibling boons) because 8534265864ec boon-renormalization
    (later in chain) drops the same columns first via
    DROP COLUMN IF EXISTS. With the fix, the alembic downgrade
    traversals past 0018 succeed cleanly until they hit
    0014_strip_account_colon which intentionally raises
    NotImplementedError (one-way design).

  - phase3_schema_changes (8b674a6a9cfc) FK on
    webhook_dlq.subscription_id is named
    webhook_dlq_subscription_id_fkey (PostgreSQL default format
    matching other FKs in models/*.py). Names aligned between
    op.create_foreign_key and op.drop_constraint so the
    downgrade FK removal works reflectively without a
    reflection-time lookup.

VERSION
  - root pyproject.toml bumped 0.12.3 -> 0.16.0 to match
    CHANGELOG [0.16.0] - 2026-07-24 entry.

VERIFIED
  - ruff check: 0 errors (apps/api/src + libs)
  - ruff format: 150 files formatted
  - mypy apps/api/src: 0 errors (67 files)
  - alembic upgrade head: 25 migrations applied end-to-end
  - alembic downgrade -1 + upgrade head: my 2 phase3 migrations
    (8b674a6a9cfc + ea3023c87c0f) are symmetrically reversible
  - alembic downgrade base: still hits 0014_strip_account_colon
    NotImplementedError by design (out of scope for this PR)

Refs: 5 ADRs in docs/adr/, .github/workflows/security.yml
(Trivy), .pre-commit-config.yaml (detect-secrets),
CONTRIBUTING.md (layered architecture), middleware.py
(RequestIDMiddleware), repositories/ (5 new files),
services/cache_service.py, services/player_service.py.


* fix(api): align roles comment + document migration chain invariants

Two small followups to the v0.16.0 release commit (3c58d5f) closing
gaps surfaced during post-release review.

1. OrmFightPlayerSummary.roles now declares
   comment="Multi-role classification: list of strings (...)"
   matching the column-level comment carried by the prod DB.
   The blank-comment in the ORM previously triggered alembic
   autogenerate to re-emit a no-op `comment=None` ALTER on every
   drift-detection pass (cosmetic DDL, no schema effect, but
   noisy changelog and unnecessary migration churn).

2. CONTRIBUTING.md gained a "Migration chain invariants (one-way
   design)" subsection documenting the two migrations that are
   intentionally not fully reversible:
   - 0014_strip_account_colon: downgrade raises NotImplementedError
     because the `:` prefix vs. bare-form distinction was lost
     on the original write (no per-row source marker survives).
   - 8b674a6a9cfc_phase3_schema_changes: all 9 destructive
     columns are restored on downgrade -1; full-chain reversal
     past this point is bounded by the 0014 boundary.
   Section includes a recovery playbook for emergency full-chain
   reversal (manual SQL operator-driven; no automatic UPDATE
   recipe is safe past 0014).

No schema-level change; no new migration emitted. Verified:
ruff clean, mypy clean on apps/api/src (67 files).

Refs: v0.16.0 release commit 3c58d5f.

---------

+++ apps/api/src/gw2analytics_api/models/upload.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from gw2analytics_api.models.fight import OrmFight
 
-from sqlalchemy import CheckConstraint, DateTime, Integer, String, Text, Uuid
+from sqlalchemy import BigInteger, CheckConstraint, DateTime, String, Text, Uuid
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from gw2analytics_api.database import Base, utcnow
@@ -30,7 +30,7 @@ class Upload(Base):
     id: Mapped[uuid.UUID] = mapped_column(Uuid, primary_key=True)
     sha256: Mapped[str] = mapped_column(String(64), unique=True)
     original_filename: Mapped[str] = mapped_column(String(255))
-    size_bytes: Mapped[int] = mapped_column(Integer)
+    size_bytes: Mapped[int] = mapped_column(BigInteger)
     uploaded_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True),
         default=utcnow,
